dashboard,api: update UI with profile picture changes (#76)
![image](https://github.com/gcivil-nyu-org/INET-Monday-Fall2023-Team-1/assets/6604943/65404d61-7443-4a24-9389-27a2a8eab8c4)

+++ furbaby/api/views.py
@@ -1,25 +1,21 @@
 import json
 import os
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponse
 from django.contrib.auth import login, logout
 from drf_standardized_errors.handler import exception_handler
 from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.generics import GenericAPIView
-from django.views.decorators.csrf import ensure_csrf_cookie
+
+# from django.views.decorators.csrf import ensure_csrf_cookie
 from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView
 from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
 from rest_framework.exceptions import PermissionDenied
-from .utils import json_response
+from .utils import json_response, make_s3_path
 from django.conf import settings
 from rest_framework.decorators import api_view
 
-from .models import Locations
-from api.auth_backends import EmailBackend
-from .models import Pets
-from .models import Locations
-from .utils import json_response
+from .models import Pets, Users, Locations
 from api.auth_backends import EmailBackend
 from .serializers import (
     RegistrationSerializer,
@@ -34,9 +30,16 @@ from django.template.loader import render_to_string
 
 from django_rest_passwordreset.signals import reset_password_token_created
 
-from rest_framework.exceptions import PermissionDenied
-import json
+from django.views.decorators.csrf import csrf_protect
 from django.core.serializers import serialize
+
+import boto3
+from botocore.exceptions import ClientError
+
+s3Config = getattr(settings, "S3_CONFIG", None)
+s3Client = boto3.client("s3", config=s3Config)
+s3AssetsFolder = getattr(settings, "ASSETS_PATH")
+s3BucketName = getattr(settings, "AWS_BUCKET_NAME")
 
 
 class UserRegistrationView(GenericAPIView):
@@ -49,6 +52,12 @@ class UserRegistrationView(GenericAPIView):
         return exception_handler
 
     def post(self, request):
+        if request.user.is_authenticated:
+            return json_response(
+                data={"message": "user account has already been created"},
+                status=status.HTTP_200_OK,
+            )
+
         serializer = self.serializer_class(data=request.data)
 
         if not serializer.is_valid():
@@ -66,10 +75,17 @@ class UserLoginView(APIView):
         return exception_handler
 
     def post(self, request):
+        if request.user.is_authenticated:
+            return json_response(
+                data={"message": "user account has already been created"},
+                status=status.HTTP_200_OK,
+            )
+
         serializer = UserLoginSerializer(data=request.data)
+
         if serializer.is_valid():
-            email = serializer.validated_data["email"]
-            password = serializer.validated_data["password"]
+            email = serializer.validated_data["email"]  # type: ignore
+            password = serializer.validated_data["password"]  # type: ignore
             email_backend = EmailBackend()
             user = email_backend.authenticate(request, email=email, password=password)
             if user is not None:
@@ -85,6 +101,7 @@ class UserLoginView(APIView):
         return json_response(data=serializer.errors, status=status.HTTP_401_UNAUTHORIZED)
 
 
+@csrf_protect
 @api_view(["GET", "OPTIONS", "POST"])
 def logout_view(request):
     if not request.user.is_authenticated:
@@ -96,14 +113,36 @@ def logout_view(request):
     return json_response({"detail": "Successfully logged out."}, status=status.HTTP_200_OK)
 
 
+@csrf_protect
 @api_view(["GET", "OPTIONS", "POST"])
 def session_view(request):
-    if not request.user.is_authenticated:
+    current_user = Users.objects.filter(id=request.user.id).first()
+
+    if not request.user.is_authenticated or current_user == None:
         return json_response({"isAuthenticated": False}, status=status.HTTP_401_UNAUTHORIZED)
 
-    return json_response({"isAuthenticated": True})
+    user_response_body = {
+        "id": current_user.id,
+        "email": current_user.email.lower(),
+    }
+
+    if current_user.first_name != None or current_user.last_name != None:
+        name = "{} {}".format(
+            "" if current_user.first_name == None else current_user.first_name,
+            "" if current_user.last_name == None else current_user.last_name,
+        )
+        user_response_body["name"] = name
+
+    return json_response(
+        {
+            "isAuthenticated": True,
+            "user": user_response_body,
+        },
+        status=status.HTTP_200_OK,
+    )
 
 
+@csrf_protect
 @api_view(["GET", "OPTIONS", "POST"])
 def whoami_view(request):
     if not request.user.is_authenticated:
@@ -112,6 +151,7 @@ def whoami_view(request):
     return json_response({"email": request.user.email}, status=status.HTTP_200_OK)
 
 
+@csrf_protect
 @api_view(["GET", "OPTIONS", "PUT", "PATCH", "DELETE"])
 def user_view(request):
     if not request.user.is_authenticated:
@@ -199,8 +239,141 @@ def password_reset_token_created(sender, instance, reset_password_token, *args, 
     msg.send()
 
 
+@csrf_protect
+@api_view(["GET", "POST", "OPTIONS"])
+def handle_profile_picture(request):
+    if not request.user.is_authenticated:
+        return json_response(
+            data={"error": "unauthenticated request. rejected"},
+            status=status.HTTP_401_UNAUTHORIZED,
+        )
+
+    if request.method == "GET":
+        return __get_user_profile_picture__(request)
+
+    if request.method == "POST":
+        return __upload_profile_picture__(request)
+
+    return json_response(
+        {
+            "error": "incorrect request method",
+            "message": 'endpoint only allows only "GET" & "PUT" requests',
+        },
+        status=status.HTTP_405_METHOD_NOT_ALLOWED,
+    )
+
+
+def __get_user_profile_picture__(request):
+    profile_picture_path = make_s3_path(
+        s3AssetsFolder, str(request.user.id), "profile-picture", "picture"
+    )
+
+    try:
+        image_object = s3Client.get_object(Bucket=s3BucketName, Key=profile_picture_path)
+        object_content = image_object["Body"].read()
+        return HttpResponse(object_content, content_type="image/jpeg")
+    except ClientError as e:
+        if e.response["Error"]["Code"] == "NoSuchKey":
+            return json_response(
+                {
+                    "data": {
+                        "message": "no profile picture present with current user: {}".format(
+                            str(request.user.id),
+                        ),
+                        "key": profile_picture_path,
+                    }
+                },
+                status=status.HTTP_404_NOT_FOUND,
+            )
+        return json_response(
+            data={
+                "error": "failed to fetch profile picture, ({})".format(e),
+                "message": "unknown error occurred while fetching user profile picture",
+            },
+            status=status.HTTP_500_INTERNAL_SERVER_ERROR,
+        )
+
+
+def __upload_profile_picture__(request):
+    if s3Config == None:
+        return json_response(
+            data={"error": "failed to upload profile picture due to internal error"},
+            status=status.HTTP_500_INTERNAL_SERVER_ERROR,
+        )
+
+    picture = request.FILES["profile_picture"]
+
+    if picture == None:
+        return json_response(
+            data={
+                "error": "file missing in upload",
+            },
+            status=status.HTTP_400_BAD_REQUEST,
+        )
+
+    if picture.content_type != "image/jpeg":
+        return json_response(
+            {
+                "error": "incorrect file format uploaded",
+                "message": 'only file type of "jpeg" is accepted',
+            },
+            status=status.HTTP_400_BAD_REQUEST,
+        )
+
+    try:
+        upload_path = make_s3_path(
+            s3AssetsFolder, str(request.user.id), "profile-picture", "picture"
+        )
+        # NOTE: if the Key is the same, the object(s) are overwritten
+        s3Client.put_object(Bucket=s3BucketName, Body=picture, Key=upload_path)
+    except ClientError as e:
+        return json_response(
+            {
+                "error": e.__str__(),
+                "message": "failed to upload file",
+            },
+            status=status.HTTP_400_BAD_REQUEST,
+        )
+
+    return json_response(
+        data={"message": "profile picture has been updated"},
+        status=status.HTTP_201_CREATED,
+    )
+
+
+# @csrf_protect
+# @api_view(["POST", "GET", "OPTIONS", "DELETE"])
+# def handle_pet_pictures(request):
+#     if not request.user.is_authenticated:
+#         return json_response(
+#             data={"error": "unauthenticated request. rejected"},
+#             status=status.HTTP_401_UNAUTHORIZED,
+#         )
+
+#     if s3Config == None:
+#         return json_response(
+#             data={"error": "failed to upload picture due to internal error"},
+#             status=status.HTTP_500_INTERNAL_SERVER_ERROR,
+#         )
+
+#     if request.method == "GET":
+#         return __get_user_pet_picture__(request)
+
+#     if request.method == "POST":
+#         return __put_user_pet_picture__(request)
+
+#     if request.method == "DELETE":
+#         return __delete_user_pet_picture__(request)
+
+#     return json_response(
+#         {
+#             "error": "incorrect request method",
+#             "message": 'endpoint only allows only "POST" & "PUT" requests',
+#         },
+
+
 # This class is for the user location(s)
-class UserLocationView(APIView):
+class UserLocationView(APIView):  # type: ignore
     # Fetch the locations serializer
     serializer_class = UserLocationSerializer
 
@@ -252,9 +425,16 @@ class UserLocationView(APIView):
 
     # takes as input a location_id and location fields and updates the location record
     def update_location_record(self, request):
+        location_id: str = request.data["id"]
+
+        if location_id == None:
+            return json_response(
+                {"message": "location_id is missing in request"},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+
         try:
             updated_fields = []
-            location_id = request.data["id"]
             location = Locations.objects.get(id=location_id)
             if "address" in request.data:
                 location.address = request.data["address"]
@@ -305,8 +485,15 @@ class UserLocationView(APIView):
             )
 
     def delete_location_record(self, request):
+        location_id: str = request.data["id"]
+
+        if location_id == None:
+            return json_response(
+                {"message": "location_id is missing in request"},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+
         try:
-            location_id = request.data["id"]
             location = Locations.objects.get(id=location_id)
             location.delete()
             return json_response(
@@ -334,36 +521,137 @@ class UserLocationView(APIView):
             )
 
 
-@api_view(["GET", "POST", "OPTIONS", "PUT", "PATCH", "DELETE"])
-def user_location_view(request):
-    location_view = UserLocationView()
+def __get_user_pet_picture__(request):
+    pet_id = request.data["pet_id"]
 
-    if not request.user.is_authenticated:
-        return json_response({"isAuthenticated": False}, status=status.HTTP_401_UNAUTHORIZED)
+    pet_info = Pets.objects.filter(name=pet_id, owner=request.user.id).first()
 
-    # fetch all user locations for the user
-    if request.method == "GET":
-        locations_list = location_view.get_user_locations(request)
+    if pet_info == None:
         return json_response(
-            locations_list, status=status.HTTP_200_OK, safe=False, include_data=False
+            data={
+                "error": "failed to locate pet with given name and user",
+            },
+            status=status.HTTP_404_NOT_FOUND,
         )
 
-    # insert a new location record for the user
-    if request.method in ["POST"]:
-        return location_view.insert_location_record(request)
+    pet_picture_path = make_s3_path(s3AssetsFolder, str(request.user.id), "pets", str(pet_info.id))
 
-    # update a location record for the user
-    if request.method in ["PUT", "PATCH"]:
-        return location_view.update_location_record(request)
+    try:
+        image_object = s3Client.get_object(Bucket=s3BucketName, Key=pet_picture_path)
+        object_content = image_object["Body"].read()
+        return HttpResponse(object_content, content_type="image/jpeg")
+    except ClientError as e:
+        if e.response["Error"]["Code"] == "NoSuchKey":
+            return json_response(
+                {
+                    "data": {
+                        "message": "no pet picture present with current user({}) and pet({})".format(
+                            request.user.id, pet_info.name
+                        )
+                    }
+                },
+                status=status.HTTP_404_NOT_FOUND,
+            )
+        return json_response(
+            data={
+                "error": "failed to fetch pet picture, ({})".format(e.__str__()),
+                "message": "unknown error occurred while fetching pet profile picture",
+            },
+            status=status.HTTP_500_INTERNAL_SERVER_ERROR,
+        )
 
-    # delete a location record for the user
-    if request.method == "DELETE":
-        return location_view.delete_location_record(request)
+
+def __put_user_pet_picture__(request):
+    pet_id = request.data["pet_id"]
+
+    pet_info = Pets.objects.filter(name=pet_id, owner=request.user.id).first()
+
+    if pet_info == None:
+        return json_response(
+            {"error": "failed to locate pet in user account"},
+            status=status.HTTP_404_NOT_FOUND,
+        )
+
+    picture = request.FILES["picture"]
+
+    if picture == None:
+        return json_response(
+            data={
+                "error": "file missing in upload",
+            },
+            status=status.HTTP_400_BAD_REQUEST,
+        )
+
+    if picture.content_type != "image/jpeg":
+        return json_response(
+            {
+                "error": "incorrect file format uploaded",
+                "message": 'only file type of "jpeg" is accepted',
+            },
+            status=status.HTTP_400_BAD_REQUEST,
+        )
+
+    try:
+        upload_path = make_s3_path(s3AssetsFolder, str(request.user.id), "pets", str(pet_info.id))
+        # NOTE: if the Key is the same, the object(s) are overwritten
+        s3Client.put_object(Bucket=s3BucketName, Body=picture, Key=upload_path)
+    except ClientError as e:
+        return json_response(
+            {
+                "error": e.__str__(),
+                "message": "failed to upload file",
+            },
+            status=status.HTTP_400_BAD_REQUEST,
+        )
 
     return json_response(
-        {"error": "incorrect request method supplied"},
-        status=status.HTTP_405_METHOD_NOT_ALLOWED,
+        data={"message": "pet picture has been updated"},
+        status=status.HTTP_201_CREATED,
     )
+
+
+def __delete_user_pet_picture__(request):
+    pet_id = request.data["pet_id"]
+
+    pet_info = Pets.objects.filter(name=pet_id, owner=request.user.id).first()
+
+    if pet_info == None:
+        return json_response(
+            data={
+                "error": "failed to locate pet with given name and user",
+            },
+            status=status.HTTP_404_NOT_FOUND,
+        )
+
+    pet_picture_path = make_s3_path(s3AssetsFolder, str(request.user.id), "pets", str(pet_info.id))
+
+    try:
+        s3Client.delete_object(Bucket=s3BucketName, Key=pet_picture_path)
+        return json_response(
+            data={
+                "data": {"message": "deleted pet({}) picture successful".format(pet_info.name)},
+            },
+            status=status.HTTP_200_OK,
+        )
+    except ClientError as e:
+        if e.response["Error"]["Code"] == "NoSuchKey":
+            return json_response(
+                {
+                    "data": {
+                        "message": "no pet picture present with current user({}) and pet({})".format(
+                            request.user.id, pet_info.name
+                        )
+                    }
+                },
+                status=status.HTTP_200_OK,
+            )
+        return json_response(
+            data={
+                "error": "failed to fetch pet picture, ({})".format(e.__str__()),
+                "message": "unknown error occurred while fetching pet picture",
+            },
+            status=status.HTTP_500_INTERNAL_SERVER_ERROR,
+        )
 
 
 class PetListCreateView(ListCreateAPIView):
@@ -373,7 +661,7 @@ class PetListCreateView(ListCreateAPIView):
 
     def get_queryset(self):
         # You can remove this method from PetRetrieveUpdateDeleteView
-        return Pets.objects.filter(owner_id=self.request.user.id)
+        return Pets.objects.filter(owner_id=self.request.user.id)  # type: ignore
 
     def create(self, request, *args, **kwargs):
         request.data["owner_id"] = request.user.id
@@ -444,9 +732,16 @@ class UserLocationView(APIView):
 
     # takes as input a location_id and location fields and updates the location record
     def update_location_record(self, request):
+        location_id: str = request.data["id"]
+
+        if location_id == None:
+            return json_response(
+                {"message": "location_id is missing in request"},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+
         try:
             updated_fields = []
-            location_id = request.data["id"]
             location = Locations.objects.get(id=location_id)
             if "address" in request.data:
                 location.address = request.data["address"]
@@ -497,8 +792,15 @@ class UserLocationView(APIView):
             )
 
     def delete_location_record(self, request):
+        location_id: str = request.data["id"]
+
+        if location_id == None:
+            return json_response(
+                {"message": "location_id is missing in request"},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+
         try:
-            location_id = request.data["id"]
             location = Locations.objects.get(id=location_id)
             location.delete()
             return json_response(
